stage2 perfected and readme comleted

+++ api/views.py
@@ -7,17 +7,8 @@ from rest_framework.response import Response
 from rest_framework import status
 
 @csrf_exempt
-@api_view(['GET'])
-def get(request, pk, *args, **kwargs):
-    """
-    Get a paerticular person from database
-    """
-    person = get_object_or_404(Person, pk=pk)
-    serializer = PersonSerializer(person, many=True)
-    return Response(serializer.data)
-
 @api_view(['POST'])
-def post(request, *args, **kwargs):
+def create(request, *args, **kwargs):
     """
     Create new person
     """
@@ -37,34 +28,33 @@ def post(request, *args, **kwargs):
             'message': 'Bad Request',
             'error': error}, status=status.HTTP_400_BAD_REQUEST)
 
-@api_view(['POST'])
-def update(request, pk, *args, **kwargs):
+@api_view(['PUT', 'GET', 'DELETE'])
+def person(request, pk, *args, **kwargs):
     """
-    Update a persons details
+    Update a persons details ie get, update or delete
     """
-    person = get_object_or_404(Person, pk=pk)
-    serialize = PersonSerializer(instance=person, serialize=request.data)
-    if serialize.is_valid():
-        serialize.save()
-        return Response({
-            'data': serialize.data,
-            'message': 'Successful'
-            }, status=status.HTTP_201_CREATED)
-
-    else:
-        default = serialize.errors
-        error = {}
-        for field_name, field_errors in default.items():
-            error[field_name] = field_errors[0]
-        return Response({
-            'message': 'Bad Request',
-            'error': error}, status=status.HTTP_400_BAD_REQUEST)
-
-@api_view(['DELETE'])
-def delete(request, pk, *args, **kwargs):
-    """
-    Delete a person from database
-    """
-    person = get_object_or_404(Person, pk=pk)
-    person.delete()
-    return Response(status=status.HTTP_202_ACCEPTED)
+    if request.method == 'PUT':
+        person = get_object_or_404(Person, pk=pk)
+        serialize = PersonSerializer(person, data=request.data)
+        if serialize.is_valid():
+            serialize.save()
+            return Response({
+                'data': serialize.data,
+                'message': 'Successful'
+                }, status=status.HTTP_201_CREATED)
+        else:
+            default = serialize.errors
+            error = {}
+            for field_name, field_errors in default.items():
+                error[field_name] = field_errors[0]
+            return Response({
+                'message': 'Bad Request',
+                'error': error}, status=status.HTTP_400_BAD_REQUEST)
+    elif request.method == 'GET':
+        person = get_object_or_404(Person, pk=pk)
+        serializer = PersonSerializer(person)
+        return Response(serializer.data, status=status.HTTP_200_OK)
+    elif request.method == 'DELETE':
+        person = get_object_or_404(Person, pk=pk)
+        person.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
